requirements.txt
sentence-transformers
qdrant-client
openai
python-dotenv
streamlit
streamlit-authenticator
pandas
torch --index-url https://download.pytorch.org/whl/cpu
torchvision --index-url https://download.pytorch.org/whl/cpu
sentence-transformers
psycopg2-binary
pyyaml
langchain
langchain-community
fastembed>=0.2.0
werkzeug
psycopg2-binary